fix: give the answer-draft retry its own cap window and keep the provider modules under the line cap
A retried draft was clamped against what the failed first attempt had already forwarded, so a long inline
reasoning pass could leave the retry a stub returned as success; the cap accountant now rebases at each
attempt while the buffer stays append-only, so the wire is bounded by two attempts of DRAFT_CAP with exactly
one retry. The budget assertion moves next to the Anthropic predicate it is sized against, the test-only
re-export goes, the extension bridge stream's tests move to their own file, and three doc comments now say
what the code does.
