+++ apps/desktop/src-tauri/src/commands/ai_provider/anthropic.rs
@@ -144,21 +144,8 @@ fn anthropic_supports_thinking(model: &str) -> bool {
         || contains_version_needle(&m, "claude-haiku-4")
 }
 
-/// Whether `max_tokens` is large enough that [`build_chat_stream_body`] turns
-/// **classic** extended thinking ON (on a model
-/// [`anthropic_supports_thinking`] accepts): Anthropic's own "big enough task
-/// to warrant it" heuristic, not a support check. Below it, no `thinking` key
-/// is sent and no budget is added on top of `max_tokens`.
-///
-/// Named (rather than left inline at the one site that branches on it)
-/// because callers OUTSIDE this module size their `max_tokens` against this
-/// threshold and have no other way to say so:
-/// `extension_bridge::answer_assist::ANSWER_ASSIST_MAX_TOKENS` is deliberately
-/// under it (that path wants LESS reasoning, and crossing it would also force
-/// `temperature` to 1.0), while `ANSWER_ASSIST_RETRY_MAX_TOKENS` deliberately
-/// crosses it. Both relationships are asserted against this function by
-/// `extension_bridge::answer_assist::tests`, so raising either budget past
-/// the gate fails a test instead of silently buying more thinking.
+/// Whether `max_tokens` reaches the gate at which [`build_chat_stream_body`]
+/// turns **classic** extended thinking on — see its call site.
 pub(crate) fn classic_thinking_engages(max_tokens: u32) -> bool {
     max_tokens >= 2048
 }
@@ -629,26 +616,19 @@ fn build_chat_stream_body(req: &AiGenerateRequest, sampling: SamplingProfile) ->
         .map(|m| json!({ "role": m.role, "content": m.content }))
         .collect();
 
-    // Thinking-token budget headroom: on BOTH classic and adaptive models,
-    // thinking tokens are billed as output tokens and count toward
-    // `max_tokens` alongside the visible response, so `max_tokens` must be
-    // inflated to leave room for both.
+    // Thinking-token budget headroom: on BOTH classic and adaptive models
+    // thinking tokens are billed as output tokens against `max_tokens`, so it
+    // must be inflated to fit them AND the visible response.
     //
-    // Classic thinking is gated on a minimum `max_tokens`
-    // (`classic_thinking_engages` — a "big enough task to warrant it"
-    // heuristic, not a support check) — below that, no `thinking` key is sent
-    // and no inflation happens (always safe; a wrongful `thinking` block 400s
-    // the whole generation on classic-only models).
-    //
-    // Adaptive thinking has NO such gate: it's on by default (several models
-    // can't even disable it) and always counts toward `max_tokens` regardless
-    // of how small the caller's budget is — a caller with a modest cap, like
-    // the extension bridge's answer-assist flow
-    // (`extension_bridge::answer_assist::ANSWER_ASSIST_MAX_TOKENS`, sized to
-    // stay UNDER the classic gate above), would otherwise get
-    // summarized-thinking tokens billed out of an un-inflated cap and come
-    // back with a short/empty draft. Reuses [`adaptive_max_tokens`] so the
-    // streaming and non-streaming builders share one formula.
+    // Classic thinking is gated on `classic_thinking_engages` (Anthropic's
+    // "big enough task" heuristic, not a support check): under the gate no
+    // `thinking` key is sent and nothing is inflated — always safe, since a
+    // wrongful `thinking` block 400s the generation on classic-only models.
+    // The gate is a named predicate because callers outside this module size
+    // budgets around it (`extension_bridge::answer_assist`: its cap under it,
+    // its one retry over — both asserted in this file's tests). Adaptive
+    // thinking has no gate at all; [`adaptive_max_tokens`] owns that case and
+    // gives the streaming and non-streaming builders one shared formula.
     let is_classic = anthropic_supports_thinking(&req.model);
     let is_adaptive = anthropic_uses_adaptive_thinking(&req.model);
     let classic_thinking_budget = if is_classic && classic_thinking_engages(max_tokens) {
